Add helper methods to check emails sent from BAT
The BAT services use a single GOV.UK Notify account for sending emails.
This commit adds a couple of helper methods to make it easy to interact
with emails that were sent as part of tested journeys.

The `wait_for_email(to:)` helper method will find the most recent email
sent to the specified address. Its intended use is during a journey
where we expect a transactional email to have been triggered by a user
action. For example, when clicking the "Sign in" button in Apply to
trigger a 'magic link' sign-in email to be sent.

Since emails are sent asynchronously, the helper will wait up to 15
seconds for the email to have been sent to Notify. This timeout is
configurable, but it means we can gracefully handle scenarios where the
app is slow to send the email – for example, due to there being a large
job queue. It'll also filter out historic emails (older than 5 seconds)
so we don't pick up emails sent from previous test runs.

The `sign_in_link` helper will extract the 'magic link' for user sign-in
from the body of the GOV.UK Notify email. This way we can visit the link
like so:

```ruby
email = wait_for_email(to: "person@example.com")
visit sign_in_link(email)
```

Note: I have updated the `.env.example` file. You'll need to provide a
NOTIFY_API_KEY with "live" permission so it can see emails sent by the
service.