Add test for signing-in to Apply with a magic link
This test demonstrates how we can use `wait_for_email` and
`sign_in_link` to handle 'magic link' sign-ins for Apply.

Note: You'll need to update your local `.env` file with the new
environment variables: `CANDIDATE_EMAIL`, `PROVIDER_ID`, `COURSE_ID`.

If this test suite were to be automated and used as a smoke test, the
`PROVIDER_ID` and `COURSE_ID` variables would ideally be populated with
the relevant ID of the newly published course (following a successful
run of `spec/publish_a_course_spec.rb`). For now I've populated the
`.env.example` file with a course which exists in the QA environment.